requirements.txt
numpy==1.9.2
scipy==0.15.1
BeautifulSoup==3.2.1
Flask==0.10.1
Flask-Script==2.0.5
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
Unirest==1.1.7
WTForms==2.1
Werkzeug==0.11.9
cffi==1.6.0
cryptography==1.3.1
dj-database-url==0.4.1
duckduckgo==0.1
enum34==1.1.4
Flask-API
flask-mongoengine==0.7.5
flask-mongo-sessions
flask_bootstrap
Flask-Session
flask-login
gnureadline==6.3.3
gunicorn==19.4.5
cchardet
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
markdown
mongoengine==0.10.6
ndg-httpsclient==0.4.0
poster==0.8.1
pandas
psycopg2==2.6.1
pyOpenSSL==16.0.0
pyasn1==0.1.9
pycparser==2.14
pymongo==3.2.2
PyJWT
requests==2.9.1
selenium
six==1.10.0
sklearn
urllib3==1.15.1
virtualenv==15.0.1
webhose==0.1.4
whitenoise==3.0
wsgiref==0.1.2
xmltodict==0.10.1
oauth2client==1.5.2
gspread
nltk